Message cleanup
git-svn-id: file:///home/jj/hercules.svn/trunk@6979 956126f8-22a0-4046-8f4a-272fa8102e63

+++ ctc_ctci.c
@@ -752,7 +752,7 @@ void  CTCI_Read( DEVBLK* pDEVBLK,   U16   sCount,
 
         if( pCTCBLK->fDebug )
         {
-            WRMSG(HHC00905, "I", SSID_TO_LCSS(pDEVBLK->ssid), pDEVBLK->devnum, iLength );
+            WRMSG(HHC00905, "I", SSID_TO_LCSS(pDEVBLK->ssid), pDEVBLK->devnum, (int)iLength );
             packet_trace( pCTCBLK->bFrameBuffer, (int)iLength, '>' );
         }
